HZN-1266: Cleanup
