+++ features/flows/classification/persistence/api/src/main/java/org/opennms/netmgt/flows/classification/persistence/api/Group.java
@@ -30,7 +30,6 @@ package org.opennms.netmgt.flows.classification.persistence.api;
 
 import java.util.ArrayList;
 import java.util.List;
-import java.util.NoSuchElementException;
 import java.util.Objects;
 
 import javax.persistence.CascadeType;
@@ -135,17 +134,5 @@ public class Group {
 
     public void setDescription(String description) {
         this.description = description;
-    }
-
-    public void replaceRuleById(Rule newRule) {
-        // Find in group
-        int ruleId = newRule.getId();
-        final Rule existingRule = getRules().stream().filter(r -> r.getId() == ruleId).findFirst().orElseThrow(() -> new NoSuchElementException("Rule with id " + ruleId + " does not exist."));
-
-        // Update values
-        existingRule.setProtocol(newRule.getProtocol());
-        existingRule.setDstAddress(newRule.getDstAddress());
-        existingRule.setName(newRule.getName());
-        existingRule.setDstPort(newRule.getDstPort());
     }
 }
