+++ features/flows/elastic/src/test/java/org/opennms/netmgt/flows/elastic/TableUtilsTest.java
@@ -31,7 +31,6 @@ package org.opennms.netmgt.flows.elastic;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.equalTo;
 
-import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collections;
 import java.util.List;
